Extract to  mapped-type.js
